create two child interactors for regularUser and VipUser
