+++ src/main/java/use_cases/UserManagerInteractor.java
@@ -48,19 +48,6 @@ public class UserManagerInteractor {
         return user.countDownRestrictionTime();
     }
 
-    public User upgrade(){
-        if (!showVIPStatus()) {
-            Upgradable upgrade_user = (Upgradable) user;
-            Hashtable<String, Object> info = upgrade_user.upgrade();
-            // TODO: Need to delete the regularUser with the same accountName first then create the VIPUser
-            UserFactory factory  = new UserFactory();
-            return factory.create(
-                    (String) info.get("password"),
-                    (String) info.get("accountName"),
-                    (Profile) info.get("profile"),
-                    true);
-        }
-        throw new IllegalArgumentException("User cannot be upgraded");
-    }
+
 
 }
